requirements.txt
discord.py
aiosqlite
dotenv
pillow